requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers>=2.5.1,<3.0.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
# groq==0.37.1
google-genai>=1.51.0
langchain-community==0.3.25
-e .